part4 15-22

+++ part4/bloglist-backend/services/blogs.js
@@ -4,9 +4,9 @@ const getAll = () => Blog.find({})
 
 const getOne = id => Blog.findOne({_id : id })
 
-const create = blog => new Blog(blog).save()
+const create = blog => new Blog({ likes : 0 ,... blog} ).save()
 
-const deleteOne  = id => Blog.findOneAndDelete({_id : id })
+const deleteOne  = (id,user_id ) => Blog.findOneAndDelete({_id : id,  user: user_id })
 
 const updateOne = (id,likes) => Blog.findOneAndUpdate({_id : id} ,{likes } ,{new:true,upsert:true ,runValidators:true})
 module.exports={getAll,create,deleteOne,getOne ,updateOne}
